fix room mention is highlighted when an expense report name has the same name
