+++ src/components/HTMLEngineProvider/HTMLRenderers/MentionReportRenderer/index.tsx
@@ -1,8 +1,6 @@
-import isEmpty from 'lodash/isEmpty';
 import React, {useContext, useMemo} from 'react';
 import type {TextStyle} from 'react-native';
 import {StyleSheet} from 'react-native';
-import type {OnyxCollection, OnyxEntry} from 'react-native-onyx';
 import {useOnyx} from 'react-native-onyx';
 import type {CustomRendererProps, TPhrasing, TText} from 'react-native-render-html';
 import {ShowContextMenuContext} from '@components/ShowContextMenuContext';
@@ -10,45 +8,17 @@ import Text from '@components/Text';
 import useCurrentReportID from '@hooks/useCurrentReportID';
 import useStyleUtils from '@hooks/useStyleUtils';
 import useThemeStyles from '@hooks/useThemeStyles';
+import {getReportMentionDetails} from '@libs/MentionUtils';
 import isSearchTopmostFullScreenRoute from '@libs/Navigation/helpers/isSearchTopmostFullScreenRoute';
 import Navigation from '@navigation/Navigation';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {Route} from '@src/ROUTES';
 import ROUTES from '@src/ROUTES';
-import type {Report} from '@src/types/onyx';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
 import MentionReportContext from './MentionReportContext';
 
 type MentionReportRendererProps = CustomRendererProps<TText | TPhrasing>;
-
-const removeLeadingLTRAndHash = (value: string) => value.replace(CONST.UNICODE.LTR, '').replace('#', '');
-
-const getMentionDetails = (htmlAttributeReportID: string, currentReport: OnyxEntry<Report>, reports: OnyxCollection<Report>, tnode: TText | TPhrasing) => {
-    let reportID: string | undefined;
-    let mentionDisplayText: string;
-
-    // Get mention details based on reportID from tag attribute
-    if (!isEmpty(htmlAttributeReportID)) {
-        const report = reports?.[`${ONYXKEYS.COLLECTION.REPORT}${htmlAttributeReportID}`];
-        reportID = report?.reportID ?? htmlAttributeReportID;
-        mentionDisplayText = removeLeadingLTRAndHash(report?.reportName ?? htmlAttributeReportID);
-        // Get mention details from name inside tnode
-    } else if ('data' in tnode && !isEmptyObject(tnode.data)) {
-        mentionDisplayText = removeLeadingLTRAndHash(tnode.data);
-
-        // eslint-disable-next-line rulesdir/prefer-early-return
-        Object.values(reports ?? {}).forEach((report) => {
-            if (report?.policyID === currentReport?.policyID && removeLeadingLTRAndHash(report?.reportName ?? '') === mentionDisplayText) {
-                reportID = report?.reportID;
-            }
-        });
-    } else {
-        return null;
-    }
-
-    return {reportID, mentionDisplayText};
-};
 
 function MentionReportRenderer({style, tnode, TDefaultRenderer, ...defaultRendererProps}: MentionReportRendererProps) {
     const styles = useThemeStyles();
@@ -65,7 +35,7 @@ function MentionReportRenderer({style, tnode, TDefaultRenderer, ...defaultRender
     // When we invite someone to a room they don't have the policy object, but we still want them to be able to see and click on report mentions, so we only check if the policyID in the report is from a workspace
     const isGroupPolicyReport = useMemo(() => currentReport && !isEmptyObject(currentReport) && !!currentReport.policyID && currentReport.policyID !== CONST.POLICY.ID_FAKE, [currentReport]);
 
-    const mentionDetails = getMentionDetails(htmlAttributeReportID, currentReport, reports, tnode);
+    const mentionDetails = getReportMentionDetails(htmlAttributeReportID, currentReport, reports, tnode);
     if (!mentionDetails) {
         return null;
     }
